* add new services for banners and slots. * refactoring

+++ internal/sdgroup/storage.go
@@ -7,5 +7,5 @@ import (
 type Storage interface {
 	Create(ctx context.Context, sdgroup SDGroup) (int64, error)
 	GetAll() ([]SDGroup, error)
-	Delete(id int) error
+	Delete(id int64) error
 }
